[refactor/#675] ExerciseConverter use-case별 mapper 분리 (#678)
* refactor: command 패키지에서 converter대신 별도 mapper를 사용하도록 변경

* refactor: 상세 조회, 게스트 조회, 참여 정보 조회 등의 책임을 Mapper로 분리

* refactor: 모임 운동 조회 관련 Mapper 분리

* refactor: 나와 관련된 조회 Mapper 분리

* refactor: 추천 관련 Mapper 분리

* refactor: converter를 없애고 map 관련 조회 로직을 Mapper로 분리

* refactor: mapper를 command/query 형태로 분리

* refactor: 거리 계산 로직을 별도 private 메서드 대신 DistanceCalculator를 사용하도록 변경

* refactor: 프로필 이미지 Url 추출 로직을 제네릭 Resolver로 해결
