+++ src/test/java/umc/cockple/demo/domain/exercise/service/ExerciseLifecycleQueryServiceTest.java
@@ -14,7 +14,8 @@ import org.springframework.data.domain.SliceImpl;
 import org.springframework.data.domain.Sort;
 import org.springframework.test.util.ReflectionTestUtils;
 import umc.cockple.demo.domain.bookmark.repository.ExerciseBookmarkRepository;
-import umc.cockple.demo.domain.exercise.converter.ExerciseConverter;
+import umc.cockple.demo.domain.exercise.converter.query.ExerciseLifecycleQueryMapper;
+import umc.cockple.demo.domain.exercise.converter.query.ExerciseParticipantInfoQueryMapper;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.exercise.domain.Guest;
 import umc.cockple.demo.domain.exercise.dto.ExerciseBuildingDetailDTO;
@@ -42,6 +43,7 @@ import umc.cockple.demo.domain.exercise.service.query.ExerciseLifecycleQueryServ
 import umc.cockple.demo.domain.member.service.support.MemberLookupService;
 import umc.cockple.demo.domain.party.service.support.PartyLookupService;
 import umc.cockple.demo.domain.file.service.FileService;
+import umc.cockple.demo.domain.file.service.ImageUrlResolver;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.member.domain.MemberExercise;
 import umc.cockple.demo.domain.member.domain.MemberParty;
@@ -99,16 +101,13 @@ class ExerciseLifecycleQueryServiceTest {
     @Mock private ExerciseBookmarkRepository exerciseBookmarkRepository;
     @Mock private FileService fileService;
 
-    private ExerciseConverter exerciseConverter;
-
     private Member manager;
     private Party party;
     private Exercise exercise;
 
     @BeforeEach
     void setUp() {
-        exerciseConverter = new ExerciseConverter(fileService);
-        exerciseLifecycleQueryService = createExerciseLifecycleQueryService(exerciseConverter);
+        exerciseLifecycleQueryService = createExerciseLifecycleQueryService();
 
         manager = MemberFixture.createMember("모임장", Gender.MALE, Level.A, 1001L);
         ReflectionTestUtils.setField(manager, "id", 1L);
@@ -123,10 +122,13 @@ class ExerciseLifecycleQueryServiceTest {
         ReflectionTestUtils.setField(exercise, "exerciseAddr", ExerciseFixture.createExerciseAddr());
     }
 
-    private ExerciseLifecycleQueryService createExerciseLifecycleQueryService(ExerciseConverter exerciseConverter) {
+    private ExerciseLifecycleQueryService createExerciseLifecycleQueryService() {
         ExerciseParticipantReader exerciseParticipantReader = new ExerciseParticipantReader(
                 memberExerciseRepository, memberPartyRepository);
         MemberLookupService memberLookupService = new MemberLookupService(memberRepository);
+        ExerciseParticipantInfoQueryMapper participantInfoMapper =
+                new ExerciseParticipantInfoQueryMapper(new ImageUrlResolver(fileService));
+        ExerciseLifecycleQueryMapper exerciseLifecycleQueryMapper = new ExerciseLifecycleQueryMapper();
 
         return new ExerciseLifecycleQueryService(
                 new ExerciseReader(exerciseRepository),
@@ -135,11 +137,11 @@ class ExerciseLifecycleQueryServiceTest {
                         exerciseParticipantReader,
                         new GuestReader(guestRepository),
                         memberLookupService,
-                        exerciseConverter
+                        participantInfoMapper
                 ),
                 memberLookupService,
                 new ExerciseValidator(memberPartyRepository, memberExerciseRepository),
-                exerciseConverter
+                exerciseLifecycleQueryMapper
         );
     }
